add ExtractionRequest.indices method and add superficial validation of ranges
Previously, there was no easy way to access the list of ranges of an ExtractionRequest
in pygribjump as no direct reference to such a Python object exists (the constructor
immediately converted the Python list of tuples to an underlying C object).
For convenience of pygribjump users, we now hold onto that reference (accessible
via a "ranges" property. For extra convenience, another "indices()" method
is added that returns a flattened 1d array of the indices of accessed values.

This commit could introduce breaking changes by validating directly in the
ExtractionRequest's constructor if a range [lo, hi) was passed where
hi <= lo, which should not be allowed. Previously, this would have only
failed at retrieval time.