simplify ExtractionRequest.ranges implementation by storing explicit copy of user-provided ranges
As pointed out in [a code review
comment](https://github.com/ecmwf/gribjump/pull/60#discussion_r2113711778), my
previous implementation (where I keep a reference to the ranges passed to an
ExtractionRequest) still had the same memory footprint as just creating a copy
of the ranges list.  Optimally, we'd provide a wrapper that allows direct
access to the underlying C array. As a short-term solution, I'll create a copy
in Python for now though.