find aec files automatically in eccodeslib via libaec_ROOT
This commit makes it possible to build gribjumplib for macos and linux by
fixing issues where libaec could not be found on macos. Part of the problem
was that the AEC_LIBRARY flag only works with filepaths and not directories.

Since FindAEC.cmake already contains the required logic to search the headers
and libraries in different folders (e.g. lib and lib64, as well as *.so and
*.dylib), we let it handle the discovery by passing the parent folder via
DAEC_LIBRARY.