Now finds libaec from libaec-config.cmake
This is in preparation to move libaec to be a source dependency in
cxx-dependencies where we will be providing libaec 1.1.4.

This is also part of bump of libaec to 1.0.6 across the whole C++ stack.
With 1.0.6 libaec introduced cmae config files that replace FindModules.
Unfortunately the cmake config file for libaec 1.0.6 to 1.1.3 do not
work correctly if libaec is not installed system wide in combination
with a system wide installed version of libaec. The bug is that the
configuration file just picks up the first libaec.h file and the first
libaec.so. Which in this combination will point to the system wide
installed version.

This can (and is addressed in this PR) by hinting cmake to the correct
location of the loaded aec module on ATOS.

When using CMAKE_PREFIX_PATH this is not an issue because those paths
will be preferred for the resolution.

Libaec 1.1.4 fixes this issue.